Convert RussianInflect class into module
