+++ lib/russian_inflect/version.rb
@@ -1,3 +1,3 @@
-class RussianInflect
+module RussianInflect
   VERSION = '1.0.2'.freeze
 end
